Moved Balance components around
