+++ packages/app-extension/src/components/Unlocked/Balances/Balances.tsx
@@ -1,0 +1,382 @@
+import {
+  Card,
+  CardContent,
+  CardHeader,
+  List,
+  ListItem,
+  ListItemIcon,
+  Typography,
+} from "@mui/material";
+import React, { useState } from "react";
+import { ExpandLess, ExpandMore } from "@mui/icons-material";
+import { formatUSD, proxyImageUrl } from "@coral-xyz/common";
+import { Button } from "../../../plugin";
+import { styles } from "@coral-xyz/themes";
+
+const useStyles = styles((theme) => ({
+  blockchainLogo: {
+    width: "12px",
+    borderRadius: "2px",
+    color: theme.custom.colors.secondary,
+  },
+  blockchainCard: {
+    backgroundColor: "inherit",
+    marginBottom: "12px",
+    marginLeft: "12px",
+    marginRight: "12px",
+    borderRadius: "12px",
+    border: theme.custom.colors.borderFull,
+  },
+  cardHeaderAvatar: {
+    display: "flex",
+    justifyContent: "center",
+    flexDirection: "column",
+  },
+  cardHeaderRoot: {
+    padding: "6px",
+    paddingLeft: "16px",
+    paddingRight: "16px",
+    height: "36px",
+    width: "100%",
+  },
+  hover: {
+    "&:hover": {
+      cursor: "pointer",
+    },
+  },
+  cardHeaderTitle: {
+    fontWeight: 500,
+    fontSize: "14px",
+  },
+  cardHeaderContent: {
+    color: theme.custom.colors.fontColor,
+  },
+  cardContentRoot: {
+    padding: "0 !important",
+  },
+  cardListRoot: {
+    padding: "0 !important",
+    height: "100%",
+  },
+  tokenListItem: {
+    borderTop: `solid 1pt ${theme.custom.colors.border}`,
+    backgroundColor: `${theme.custom.colors.nav} !important`,
+    paddingTop: 0,
+    paddingBottom: 0,
+    paddingLeft: "12px",
+    paddingRight: "12px",
+    padding: 0,
+    height: "68px",
+  },
+  balancesTableCellContainer: {
+    width: "100%",
+    height: "100%",
+    display: "flex",
+  },
+  tokenListItemContent: {
+    color: theme.custom.colors.fontColor,
+    flex: 1,
+    paddingTop: "10px",
+    paddingBottom: "10px",
+  },
+  tokenListItemRow: {
+    display: "flex",
+    justifyContent: "space-between",
+  },
+  logoIcon: {
+    borderRadius: "22px",
+    width: "44px",
+    height: "44px",
+  },
+  tokenListItemIcon: {
+    paddingTop: "12px",
+    paddingBottom: "12px",
+    marginRight: "12px",
+  },
+  tokenName: {
+    height: "24px",
+    fontWeight: 500,
+    fontSize: "16px",
+    maxWidth: "200px",
+    overflow: "hidden",
+    color: theme.custom.colors.fontColor,
+    lineHeight: "24px",
+  },
+  tokenAmount: {
+    fontWeight: 500,
+    fontSize: "14px",
+    color: theme.custom.colors.secondary,
+    lineHeight: "20px",
+  },
+  tokenBalance: {
+    fontWeight: 500,
+    fontSize: "16px",
+    color: theme.custom.colors.fontColor,
+    lineHeight: "24px",
+  },
+  tokenBalanceChangeNeutral: {
+    fontWeight: 500,
+    fontSize: "14px",
+    color: theme.custom.colors.secondary,
+    float: "right",
+    lineHeight: "20px",
+  },
+  tokenBalanceChangePositive: {
+    fontWeight: 500,
+    fontSize: "14px",
+    color: theme.custom.colors.positive,
+    float: "right",
+    lineHeight: "20px",
+  },
+  tokenBalanceChangeNegative: {
+    fontWeight: 500,
+    fontSize: "12px",
+    color: theme.custom.colors.negative,
+    float: "right",
+  },
+  tokenListItemIconRoot: {
+    minWidth: "44px",
+  },
+  expand: {
+    width: "18px",
+    color: theme.custom.colors.secondary,
+  },
+}));
+
+export function BalancesTableCell({ props, style }: any) {
+  const { icon, title, subtitle, usdValue, percentChange } = props;
+  const classes = useStyles();
+
+  const positive = percentChange && percentChange > 0 ? true : false;
+  const negative = percentChange && percentChange < 0 ? true : false;
+  const neutral = percentChange && percentChange === 0 ? true : false;
+
+  return (
+    <div className={classes.balancesTableCellContainer}>
+      {!!icon && (
+        <ListItemIcon
+          className={classes.tokenListItemIcon}
+          classes={{ root: classes.tokenListItemIconRoot }}
+        >
+          <ProxyImage
+            src={icon}
+            className={classes.logoIcon}
+            onError={(event: any) =>
+              (event.currentTarget.style.display = "none")
+            }
+          />
+        </ListItemIcon>
+      )}
+      <div className={classes.tokenListItemContent}>
+        <div className={classes.tokenListItemRow}>
+          <Typography className={classes.tokenName}>{title}</Typography>
+          {usdValue && (
+            <Typography className={classes.tokenBalance}>
+              {formatUSD(usdValue)}
+            </Typography>
+          )}
+        </div>
+        <div className={classes.tokenListItemRow}>
+          {subtitle && (
+            <Typography className={classes.tokenAmount}>{subtitle}</Typography>
+          )}
+          {percentChange !== undefined && positive && (
+            <Typography className={classes.tokenBalanceChangePositive}>
+              +{formatUSD(percentChange.toLocaleString())}
+            </Typography>
+          )}
+          {percentChange !== undefined && negative && (
+            <Typography className={classes.tokenBalanceChangeNegative}>
+              {formatUSD(percentChange.toLocaleString())}
+            </Typography>
+          )}
+          {percentChange !== undefined && neutral && (
+            <Typography className={classes.tokenBalanceChangeNeutral}>
+              {formatUSD(percentChange.toLocaleString())}
+            </Typography>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export function BalancesTable({ props, style, children }: any) {
+  const classes = useStyles();
+  return (
+    <BalancesTableProvider>
+      <Card className={classes.blockchainCard} elevation={0} style={style}>
+        {children}
+      </Card>
+    </BalancesTableProvider>
+  );
+}
+
+function BalancesTableProvider(props: any) {
+  const [showContent, setShowContent] = useState(true);
+  return (
+    <_BalancesTableContext.Provider
+      value={{
+        showContent,
+        setShowContent,
+      }}
+    >
+      {props.children}
+    </_BalancesTableContext.Provider>
+  );
+}
+
+type BalancesContext = {
+  showContent: boolean;
+  setShowContent: (b: boolean) => void;
+};
+const _BalancesTableContext = React.createContext<BalancesContext | null>(null);
+
+export function useBalancesContext() {
+  const ctx = React.useContext(_BalancesTableContext);
+  if (ctx === null) {
+    throw new Error("Context not available");
+  }
+  return ctx;
+}
+
+export function BalancesTableHead({ props, style }: any) {
+  const { subtitle, title, iconUrl, disableToggle } = props;
+  const classes = useStyles();
+  const { showContent, setShowContent } = useBalancesContext();
+  return (
+    <Button
+      style={{
+        width: "100%",
+        borderRadius: 0,
+        padding: 0,
+        ...style,
+      }}
+    >
+      <CardHeader
+        onClick={() => !disableToggle && setShowContent(!showContent)}
+        avatar={
+          iconUrl ? (
+            <ProxyImage className={classes.blockchainLogo} src={iconUrl} />
+          ) : undefined
+        }
+        title={
+          <div
+            style={{
+              display: "flex",
+              justifyContent: "space-between",
+            }}
+          >
+            <div
+              style={{
+                display: "flex",
+              }}
+            >
+              <Typography
+                style={{
+                  fontWeight: 500,
+                  lineHeight: "24px",
+                  fontSize: "14px",
+                  display: "flex",
+                  justifyContent: "center",
+                  flexDirection: "column",
+                }}
+              >
+                {title}
+              </Typography>
+              {subtitle}
+            </div>
+            {!disableToggle && (
+              <>
+                {showContent ? (
+                  <ExpandLess className={classes.expand} />
+                ) : (
+                  <ExpandMore className={classes.expand} />
+                )}
+              </>
+            )}
+          </div>
+        }
+        classes={{
+          root: `${classes.cardHeaderRoot} ${
+            disableToggle ? "" : classes.hover
+          }`,
+          content: classes.cardHeaderContent,
+          title: classes.cardHeaderTitle,
+          avatar: classes.cardHeaderAvatar,
+        }}
+      />
+    </Button>
+  );
+}
+
+export function BalancesTableContent({ props, style, children }: any) {
+  const classes = useStyles();
+  const { showContent } = useBalancesContext();
+  return (
+    <CardContent classes={{ root: classes.cardContentRoot }} style={style}>
+      <List
+        style={{
+          display: !showContent ? "none" : undefined,
+        }}
+        classes={{ root: classes.cardListRoot }}
+      >
+        {children}
+      </List>
+    </CardContent>
+  );
+}
+
+export function BalancesTableRow({
+  id,
+  props,
+  style,
+  children,
+  childrenRenderer,
+  onClick,
+}: any) {
+  return (
+    <__BalancesTableRow
+      id={id}
+      props={props}
+      style={style}
+      children={children}
+      childrenRenderer={childrenRenderer}
+      onClick={onClick}
+    />
+  );
+}
+
+function __BalancesTableRow({
+  id,
+  props,
+  style,
+  children,
+  childrenRenderer,
+  onClick,
+}: any) {
+  const classes = useStyles();
+  return (
+    <ListItem
+      button
+      disableRipple
+      className={classes.tokenListItem}
+      onClick={onClick}
+      style={style}
+    >
+      {children}
+    </ListItem>
+  );
+}
+function ProxyImage(props: any) {
+  return (
+    <img
+      {...props}
+      onError={({ currentTarget }) => {
+        currentTarget.onerror = props.onError || null;
+        currentTarget.src = props.src;
+      }}
+      src={proxyImageUrl(props.src)}
+    />
+  );
+}
